% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rwbtw}
\alias{.rwbtw}
\title{Random-walk Betweenness}
\usage{
.rwbtw(A)
}
\arguments{
\item{A}{(weighted) adjacency matrix}
}
\value{
returns the random walk betweenness of each node
}
\description{
Random-walk Betweenness
}
