% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-data.R
\docType{data}
\name{network1}
\alias{network1}
\alias{network2}
\title{Example networks}
\format{
Each is a `matrix` object with 1 representing edges between the nodes
}
\usage{
network1

network2
}
\description{
These matrices are the adjacency matrices used in Figure 3 of Newman(2005)
}
\details{
`network1` is the adjacency matrix of the network on the right of the figure (Network 1) and `network2` is the adjacency matrix of the network on the left of the figure (Network 2)
}
\section{References}{
 Newman, Mark EJ. "A measure of betweenness centrality based on random walks." Social networks 27, no. 1 (2005): 39-54.
}

\examples{
network1
network2
}
\keyword{datasets}
