% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwbtw-package.R
\docType{package}
\name{rwbtw-package}
\alias{rwbtw-package}
\title{rwbtw: random walk betweenness}
\description{
Package to calculate the random walk betweenness score of nodes in (possibly weighted) graphs
}
\keyword{internal}
