% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwbtw.R
\name{rwbtw}
\alias{rwbtw}
\title{Random walk betweenness}
\usage{
rwbtw(x)
}
\arguments{
\item{x}{a symemtric (possibly weighted) adjacency matrix or an \code{igraph} object}
}
\value{
returns the random walk betweenness
}
\description{
Calculates the random walk betweenness of each node in a graph
}
\details{
if \code{x} is an \code{igraph} object that contains an edge attribute named \code{weight}, it will be used by default
}
\section{References}{
 Newman, Mark EJ. "A measure of betweenness centrality based on random walks." Social networks 27, no. 1 (2005): 39-54.
}

